+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - EntityQueryModel.resetConditionChanged() removed.
 - EntitySearchModel.Builder.condition() added.
 - ForeignKeyConditionModel.equalSearchModel() and inSearchModel() moved to ForeignKeyOperands.
+- EntityConditionModel.attribute() replaced with column() and foreignKey().
 ### is.codion.swing.framework.model
 - SwingForeignKeyConditionModel.equalComboBoxModel() and inSearchModel() moved to SwingForeignKeyOperands.
 ### is.codion.swing.framework.ui